# backend/inventory/api/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package http

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"jolly/backend/common"

	"github.com/labstack/echo/v4"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// StockResponse defines model for StockResponse.
type StockResponse struct {
	Quantity  int         `json:"quantity"`
	Sku       string      `json:"sku"`
	StockUuid common.UUID `json:"stock_uuid"`
}

// UpsertStockRequest defines model for UpsertStockRequest.
type UpsertStockRequest struct {
	Quantity int    `json:"quantity"`
	Sku      string `json:"sku"`
}

// UpsertStockJSONRequestBody defines body for UpsertStock for application/json ContentType.
type UpsertStockJSONRequestBody = UpsertStockRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create or update inventory stock for a variant SKU
	// (POST /inventory/stock)
	UpsertStock(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// UpsertStock converts echo context to params.
func (w *ServerInterfaceWrapper) UpsertStock(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpsertStock(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/inventory/stock", wrapper.UpsertStock)

}

type UpsertStockRequestObject struct {
	Body *UpsertStockJSONRequestBody
}

type UpsertStockResponseObject interface {
	VisitUpsertStockResponse(w http.ResponseWriter) error
}

type UpsertStock200JSONResponse StockResponse

func (response UpsertStock200JSONResponse) VisitUpsertStockResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type UpsertStock400JSONResponse ErrorResponse

func (response UpsertStock400JSONResponse) VisitUpsertStockResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Create or update inventory stock for a variant SKU
	// (POST /inventory/stock)
	UpsertStock(ctx context.Context, request UpsertStockRequestObject) (UpsertStockResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// UpsertStock operation middleware
func (sh *strictHandler) UpsertStock(ctx echo.Context) error {
	var request UpsertStockRequestObject

	var body UpsertStockJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.UpsertStock(ctx.Request().Context(), request.(UpsertStockRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "UpsertStock")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(UpsertStockResponseObject); ok {
		return validResponse.VisitUpsertStockResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}