# backend/orders/api/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package http

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"jolly/backend/common/shared"
	"jolly/backend/orders/app"
	"jolly/backend/orders/domain"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// Defines values for OrderStatus.
const (
	Confirmed      OrderStatus = "confirmed"
	Failed         OrderStatus = "failed"
	Paid           OrderStatus = "paid"
	PendingPayment OrderStatus = "pending_payment"
)

// CreateOrder defines model for CreateOrder.
type CreateOrder struct {
	Currency   Currency          `json:"currency"`
	CustomerId string            `json:"customer_id"`
	Items      []CreateOrderItem `json:"items"`
}

// CreateOrderItem defines model for CreateOrderItem.
type CreateOrderItem struct {
	Quantity       int    `json:"quantity"`
	Sku            string `json:"sku"`
	UnitPriceCents int64  `json:"unit_price_cents"`
}

// Currency defines model for Currency.
type Currency = shared.Currency

// ErrorDetail defines model for ErrorDetail.
type ErrorDetail struct {
	EntityId   *string `json:"entity_id,omitempty"`
	EntityType *string `json:"entity_type,omitempty"`
	ErrorSlug  *string `json:"error_slug,omitempty"`
	Message    *string `json:"message,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Details *[]ErrorDetail `json:"details,omitempty"`
	Message string         `json:"message"`
	Slug    string         `json:"slug"`
}

// LineItemUUID Unique line item identifier
type LineItemUUID = domain.LineItemUUID

// Order defines model for Order.
type Order struct {
	CreatedAt  time.Time       `json:"created_at"`
	Currency   Currency        `json:"currency"`
	CustomerId string          `json:"customer_id"`
	Items      []OrderLineItem `json:"items"`

	// OrderUuid Unique order identifier
	OrderUuid   OrderUUID   `json:"order_uuid"`
	PlacedAtUtc time.Time   `json:"placed_at_utc"`
	Status      OrderStatus `json:"status"`
	TotalCents  int64       `json:"total_cents"`
	UpdatedAt   time.Time   `json:"updated_at"`
}

// OrderCreated defines model for OrderCreated.
type OrderCreated struct {
	// OrderUuid Unique order identifier
	OrderUuid OrderUUID   `json:"order_uuid"`
	Status    OrderStatus `json:"status"`
}

// OrderLineItem defines model for OrderLineItem.
type OrderLineItem struct {
	// LineItemUuid Unique line item identifier
	LineItemUuid   LineItemUUID `json:"line_item_uuid"`
	Quantity       int32        `json:"quantity"`
	Sku            string       `json:"sku"`
	UnitPriceCents int64        `json:"unit_price_cents"`
}

// OrderStatus defines model for OrderStatus.
type OrderStatus string

// OrderUUID Unique order identifier
type OrderUUID = app.OrderUUID

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// Unauthorized defines model for Unauthorized.
type Unauthorized = ErrorResponse

// CreateOrderJSONRequestBody defines body for CreateOrder for application/json ContentType.
type CreateOrderJSONRequestBody = CreateOrder

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create a new order
	// (POST /orders)
	CreateOrder(ctx echo.Context) error
	// Get an order by UUID
	// (GET /orders/{order_uuid})
	GetOrder(ctx echo.Context, orderUuid OrderUUID) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// CreateOrder converts echo context to params.
func (w *ServerInterfaceWrapper) CreateOrder(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateOrder(ctx)
	return err
}

// GetOrder converts echo context to params.
func (w *ServerInterfaceWrapper) GetOrder(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "order_uuid" -------------
	var orderUuid OrderUUID

	err = runtime.BindStyledParameterWithOptions("simple", "order_uuid", ctx.Param("order_uuid"), &orderUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter order_uuid: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetOrder(ctx, orderUuid)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/orders", wrapper.CreateOrder)
	router.GET(baseURL+"/orders/:order_uuid", wrapper.GetOrder)

}

type BadRequestJSONResponse ErrorResponse

type NotFoundJSONResponse ErrorResponse

type UnauthorizedJSONResponse ErrorResponse

type CreateOrderRequestObject struct {
	Body *CreateOrderJSONRequestBody
}

type CreateOrderResponseObject interface {
	VisitCreateOrderResponse(w http.ResponseWriter) error
}

type CreateOrder201JSONResponse OrderCreated

func (response CreateOrder201JSONResponse) VisitCreateOrderResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type CreateOrder400JSONResponse struct{ BadRequestJSONResponse }

func (response CreateOrder400JSONResponse) VisitCreateOrderResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type CreateOrder401JSONResponse struct{ UnauthorizedJSONResponse }

func (response CreateOrder401JSONResponse) VisitCreateOrderResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type GetOrderRequestObject struct {
	OrderUuid OrderUUID `json:"order_uuid"`
}

type GetOrderResponseObject interface {
	VisitGetOrderResponse(w http.ResponseWriter) error
}

type GetOrder200JSONResponse Order

func (response GetOrder200JSONResponse) VisitGetOrderResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetOrder400JSONResponse struct{ BadRequestJSONResponse }

func (response GetOrder400JSONResponse) VisitGetOrderResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetOrder401JSONResponse struct{ UnauthorizedJSONResponse }

func (response GetOrder401JSONResponse) VisitGetOrderResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type GetOrder404JSONResponse struct{ NotFoundJSONResponse }

func (response GetOrder404JSONResponse) VisitGetOrderResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Create a new order
	// (POST /orders)
	CreateOrder(ctx context.Context, request CreateOrderRequestObject) (CreateOrderResponseObject, error)
	// Get an order by UUID
	// (GET /orders/{order_uuid})
	GetOrder(ctx context.Context, request GetOrderRequestObject) (GetOrderResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// CreateOrder operation middleware
func (sh *strictHandler) CreateOrder(ctx echo.Context) error {
	var request CreateOrderRequestObject

	var body CreateOrderJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.CreateOrder(ctx.Request().Context(), request.(CreateOrderRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateOrder")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(CreateOrderResponseObject); ok {
		return validResponse.VisitCreateOrderResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetOrder operation middleware
func (sh *strictHandler) GetOrder(ctx echo.Context, orderUuid OrderUUID) error {
	var request GetOrderRequestObject

	request.OrderUuid = orderUuid

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetOrder(ctx.Request().Context(), request.(GetOrderRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetOrder")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetOrderResponseObject); ok {
		return validResponse.VisitGetOrderResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}