# backend/orders/api/http/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Orders
  contact:
    name: jolly
paths:
  /orders:
    post:
      summary: Create a new order
      operationId: createOrder
      tags:
        - orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrder"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /orders/{order_uuid}:
    get:
      summary: Get an order by UUID
      operationId: getOrder
      tags:
        - orders
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OrderUUID"
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  schemas:
    CreateOrder:
      type: object
      required:
        - customer_id
        - currency
        - items
      properties:
        customer_id:
          type: string
        currency:
          $ref: "#/components/schemas/Currency"
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateOrderItem"

    CreateOrderItem:
      type: object
      required:
        - sku
        - quantity
        - unit_price_cents
      properties:
        sku:
          type: string
        quantity:
          type: integer
          minimum: 1
        unit_price_cents:
          type: integer
          format: int64
          minimum: 0

    OrderCreated:
      type: object
      required:
        - order_uuid
        - status
      properties:
        order_uuid:
          $ref: "#/components/schemas/OrderUUID"
        status:
          $ref: "#/components/schemas/OrderStatus"

    Order:
      type: object
      required:
        - order_uuid
        - customer_id
        - currency
        - status
        - total_cents
        - placed_at_utc
        - created_at
        - updated_at
        - items
      properties:
        order_uuid:
          $ref: "#/components/schemas/OrderUUID"
        customer_id:
          type: string
        currency:
          $ref: "#/components/schemas/Currency"
        status:
          $ref: "#/components/schemas/OrderStatus"
        total_cents:
          type: integer
          format: int64
        placed_at_utc:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"

    OrderLineItem:
      type: object
      required:
        - line_item_uuid
        - sku
        - quantity
        - unit_price_cents
      properties:
        line_item_uuid:
          $ref: "#/components/schemas/LineItemUUID"
        sku:
          type: string
        quantity:
          type: integer
          format: int32
        unit_price_cents:
          type: integer
          format: int64

    OrderUUID:
      type: string
      format: uuid
      description: Unique order identifier
      x-go-type: app.OrderUUID
      x-go-type-import:
        path: jolly/backend/orders/app

    LineItemUUID:
      type: string
      format: uuid
      description: Unique line item identifier
      x-go-type: domain.LineItemUUID
      x-go-type-import:
        path: jolly/backend/orders/domain

    OrderStatus:
      type: string
      enum:
        - pending_payment
        - paid
        - confirmed
        - failed

    Currency:
      type: string
      x-go-type: shared.Currency
      x-go-type-import:
        path: jolly/backend/common/shared

    ErrorResponse:
      type: object
      required:
        - message
        - slug
      properties:
        message:
          type: string
        slug:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"

    ErrorDetail:
      type: object
      properties:
        entity_type:
          type: string
        entity_id:
          type: string
        error_slug:
          type: string
        message:
          type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"